+++ front/lib/iam/invitations.ts
@@ -9,7 +9,8 @@ import { verify } from "jsonwebtoken";
 import config from "@app/lib/api/config";
 import { getMembershipInvitationUrl } from "@app/lib/api/invitation";
 import { AuthFlowError } from "@app/lib/iam/errors";
-import { MembershipInvitation, Workspace } from "@app/lib/models/workspace";
+import { MembershipInvitation } from "@app/lib/models/membership_invitation";
+import { Workspace } from "@app/lib/models/workspace";
 import type { UserResource } from "@app/lib/resources/user_resource";
 import { renderLightWorkspaceType } from "@app/lib/workspace";
 import logger from "@app/logger/logger";
