+++ front/lib/resources/types.ts
@@ -7,10 +7,8 @@ import type {
   WhereOptions,
 } from "sequelize";
 
-import type {
-  BaseModel,
-  SoftDeletableModel,
-} from "@app/lib/resources/storage/wrappers/base";
+import type { BaseModel } from "@app/lib/resources/storage/wrappers/base";
+import type { WorkspaceSoftDeletableModel } from "@app/lib/resources/storage/wrappers/model_with_workspace";
 
 export type NonAttributeKeys<M> = {
   [K in keyof M]: M[K] extends NonAttribute<infer T>
@@ -51,6 +49,6 @@ export type ResourceFindOptions<M extends Model> = {
   limit?: number;
   order?: FindOptions<M>["order"];
   where?: WhereOptions<M>;
-} & (M extends SoftDeletableModel
+} & (M extends WorkspaceSoftDeletableModel
   ? { includeDeleted?: boolean }
   : { includeDeleted?: never });
