+++ front/pages/api/w/[wId]/dust_app_secrets/index.ts
@@ -8,7 +8,7 @@ import {
   getDustAppSecrets,
 } from "@app/lib/api/dust_app_secrets";
 import type { Authenticator } from "@app/lib/auth";
-import { DustAppSecret } from "@app/lib/models/workspace";
+import { DustAppSecret } from "@app/lib/models/dust_app_secret";
 import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
